move server/client configuration for tests to separate unit
