+++ streamer/Cargo.toml
@@ -29,6 +29,7 @@ quinn-proto = { workspace = true }
 rand = { workspace = true }
 rustls = { workspace = true, features = ["dangerous_configuration"] }
 smallvec = { workspace = true }
+socket2 = { workspace = true }
 solana-measure = { workspace = true }
 solana-metrics = { workspace = true }
 solana-perf = { workspace = true }
@@ -40,7 +41,6 @@ x509-parser = { workspace = true }
 
 [dev-dependencies]
 assert_matches = { workspace = true }
-socket2 = { workspace = true }
 solana-logger = { workspace = true }
 
 [lib]
