+++ streamer/src/nonblocking/mod.rs
@@ -5,3 +5,4 @@ pub mod rate_limiter;
 pub mod recvmmsg;
 pub mod sendmmsg;
 mod stream_throttle;
+pub mod testing_utilities;
